[Android] Suggest components moved to appropriate packages.
Chromium change:

https://source.chromium.org/chromium/chromium/src/+/6e3695055282bde2999201b8c5a7629428831f2e

commit 6e3695055282bde2999201b8c5a7629428831f2e
Author: Tomasz Wiszkowski <ender@google.com>
Date:   Fri Aug 4 17:03:26 2023 +0000

    Move suggest components to appropriate packages.

    The change moves
    - SuggestionSpanable to styles package (away from `base`),
    - HistoryClustersProcessor to history_clusters package (ditto).

    The change bumps test coverage for both relocated classes to 100%.
